Add recent changes on master to release (#185)
* Parse from db to graphQL (#157)

* Query from db and put into GraphQL schema format

* Make master use db & create cron job

* Fix saving & parsing of dining items

* Add installation for cron in Dockerfile (#167)

* Fix parsing and DB storage of operating hours (#169)

* Add macs to static slugs (#170)

* Fix missing swipe data for Terrace (#171)

* Update cron job to run daily at 4am, 10am, 4pm (#174)

* Add crontab log to Dockerfile (#175)

* Add newline to cronjob txt file

* Fix static eatery category (#176)

* Fix date to use Eastern timezone (#177)

* January hours for Macs, Terrace, and Manndibles (#179)

Will need an update towards the end of January - Manndibles hours change

* Support new BRB accounts (#181)

* Change script field and add to SSH host for Travis (#183)

* Update externalEateries.json (#180)

* Add conditional statement to update collegetown eateries (#178)

* Fix cronjob to run in container (#184)

* Remove trailing comma

* Final fix to get campus and ctown cronjobs working (#186)

* Clean up unused code (#188)

* Update externalEateries.json (#187)

* Fix formatting in externalEateries.json (#189)

+++ src/db.py
@@ -3,7 +3,7 @@ import requests
 
 from .collegetown import collegetown_search
 from .constants import CORNELL_DINING_URL, STATIC_EATERIES_URL, STATIC_EATERY_SLUGS, STATIC_EXPANDED_ITEMS_URL
-from .database import CampusEatery, CollegetownEatery, CollegetownEateryHour, Base, Engine, Session, SwipeData
+from .database import Base, CampusEatery, CollegetownEatery, CollegetownEateryHour, Engine, Session, SwipeData
 from .eatery_db import (
     export_data,
     parse_campus_eateries,
@@ -20,10 +20,10 @@ from .eatery_db import (
     parse_to_csv,
 )
 
+conn = Engine.connect()
+
 
 def get_campus_eateries(data_json, refresh=False):
-    campus_eateries = CampusEatery.query.all()
-
     if refresh:
         Base.metadata.drop_all(bind=Engine)
         Base.metadata.create_all(bind=Engine)
@@ -31,24 +31,17 @@ def get_campus_eateries(data_json, refresh=False):
         campus_eateries = parse_campus_eateries(data_json)
         Session.add_all(campus_eateries)
         Session.commit()
+    else:
+        campus_eateries = CampusEatery.query.all()
 
     return campus_eateries
 
 
-def start_update(refresh_campus=False, recalculate_swipe=False):
+def start_update(refresh_campus=False, recalculate_swipe=False, refresh_collegetown=False):
     try:
         print("[{}] Fetching campus eateries".format(datetime.now()))
         campus_json = requests.get(CORNELL_DINING_URL).json()
         campus_eateries = get_campus_eateries(campus_json, refresh=refresh_campus)
-
-        if recalculate_swipe:
-            print("[{}] Updating swipe data".format(datetime.now()))
-            data_path = parse_to_csv(file_name="data.csv")
-            Base.metadata.drop_all(bind=Engine, tables=[SwipeData.__table__])
-            Base.metadata.create_all(bind=Engine, tables=[SwipeData.__table__])
-            all_swipe_data = export_data(data_path, campus_eateries)
-            Session.add_all(all_swipe_data)
-            Session.commit()
 
         print("[{}] Updating campus eatery hours and menus".format(datetime.now()))
         for eatery in campus_eateries:
@@ -72,13 +65,17 @@ def start_update(refresh_campus=False, recalculate_swipe=False):
 
             # if this is not a refresh, then static eateries are part of campus_eateries
             if eatery.slug not in STATIC_EATERY_SLUGS:
-                hours_and_menus = parse_campus_hours(campus_json, eatery)
+                hours_and_menus, dining_items = parse_campus_hours(campus_json, eatery)
                 eatery_hours = (x[0] for x in hours_and_menus)
                 Session.add_all(eatery_hours)
                 Session.commit()
 
+                if dining_items:
+                    hours_and_menus.append((None, dining_items))
+
                 for eatery_hour, menu_json in hours_and_menus:
-                    categories_and_items = parse_menu_categories(menu_json, eatery_hour)
+                    categories_and_items = parse_menu_categories(menu_json, eatery_hour, eatery.id)
+
                     eatery_categories = (x[0] for x in categories_and_items)
                     Session.add_all(eatery_categories)
                     Session.commit()
@@ -107,7 +104,7 @@ def start_update(refresh_campus=False, recalculate_swipe=False):
                 Session.commit()
 
                 for eatery_hour, menu_json in hours_and_menus:
-                    categories_and_items = parse_menu_categories(menu_json, eatery_hour)
+                    categories_and_items = parse_menu_categories(menu_json, eatery_hour, eatery.id)
                     eatery_categories = (x[0] for x in categories_and_items)
                     Session.add_all(eatery_categories)
                     Session.commit()
@@ -117,19 +114,29 @@ def start_update(refresh_campus=False, recalculate_swipe=False):
                         Session.add_all(menu_items)
                         Session.commit()
 
-        Base.metadata.drop_all(bind=Engine, tables=[CollegetownEatery.__table__, CollegetownEateryHour.__table__])
-        Base.metadata.create_all(bind=Engine, tables=[CollegetownEatery.__table__, CollegetownEateryHour.__table__])
-        print("[{}] Fetching Collegetown eateries".format(datetime.now()))
-        yelp_query = collegetown_search()
-        collegetown_eateries = parse_collegetown_eateries(yelp_query)
-        Session.add_all(collegetown_eateries)
-        Session.commit()
-
-        print("[{}] Updating Collegetown eateries and hours".format(datetime.now()))
-        for eatery in collegetown_eateries:
-            hours = parse_collegetown_hours(yelp_query, eatery)
-            Session.add_all(hours)
+        if recalculate_swipe:
+            print("[{}] Updating swipe data".format(datetime.now()))
+            data_path = parse_to_csv(file_name="data.csv")
+            Base.metadata.drop_all(bind=Engine, tables=[SwipeData.__table__])
+            Base.metadata.create_all(bind=Engine, tables=[SwipeData.__table__])
+            all_swipe_data = export_data(data_path, campus_eateries + static_eateries)
+            Session.add_all(all_swipe_data)
             Session.commit()
+
+        if refresh_collegetown:
+            Base.metadata.drop_all(bind=Engine, tables=[CollegetownEatery.__table__, CollegetownEateryHour.__table__])
+            Base.metadata.create_all(bind=Engine, tables=[CollegetownEatery.__table__, CollegetownEateryHour.__table__])
+            print("[{}] Fetching Collegetown eateries".format(datetime.now()))
+            yelp_query = collegetown_search()
+            collegetown_eateries = parse_collegetown_eateries(yelp_query)
+            Session.add_all(collegetown_eateries)
+            Session.commit()
+
+            print("[{}] Updating Collegetown eateries and hours".format(datetime.now()))
+            for eatery in collegetown_eateries:
+                hours = parse_collegetown_hours(yelp_query, eatery)
+                Session.add_all(hours)
+                Session.commit()
 
     except Exception as e:
         print("Data update failed:", e)
